+++ ricaun.Revit.UI.Example/Revit/App.cs
@@ -2,6 +2,7 @@ using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using ricaun.Revit.UI;
+using ricaun.Revit.UI.Drawing;
 using ricaun.Revit.UI.Example.Proprieties;
 using ricaun.Revit.UI.Utils;
 using System;